Implement `ValidationCheck` for check
